+++ renderapi/transform/leaf/__init__.py
@@ -1,5 +1,5 @@
 from .transform import *
-from .AffineModels import *
-from .PolynomialModels import *
-from .ThinPlateSpline import *
+from .affine_models import *
+from .polynomial_models import *
+from .thin_plate_spline import *
 from .utils import *
